Integerize position handling

+++ testmud/mud/home/Game/lib/thing/position.c
@@ -23,7 +23,7 @@
 
 inherit "~Thing/lib/thing";
 
-private float xpos, ypos, zpos;
+private int xpos, ypos, zpos;
 
 /*********************/
 /* Position handling */
@@ -59,32 +59,32 @@ static void move_notify(object old_env)
 	}
 }
 
-void set_x_position(float new_xpos)
+void set_x_position(int new_xpos)
 {
 	xpos = new_xpos;
 }
 
-void set_y_position(float new_ypos)
+void set_y_position(int new_ypos)
 {
 	ypos = new_ypos;
 }
 
-void set_z_position(float new_zpos)
+void set_z_position(int new_zpos)
 {
 	zpos = new_zpos;
 }
 
-float query_x_position()
+int query_x_position()
 {
 	return xpos;
 }
 
-float query_y_position()
+int query_y_position()
 {
 	return ypos;
 }
 
-float query_z_position()
+int query_z_position()
 {
 	return zpos;
 }
